Split code in index.js to multiple modules
