+++ modules/comments-popup.js
@@ -1,17 +1,23 @@
-function createCommentsPopup(showObj) {
+export default function createCommentsPopup(showObj) {
   const popupWindow = document.createElement('div');
+  const closeBtt = document.createElement('div');
+  closeBtt.classList.add('close-butt');
+  popupWindow.id = 'comments-popup';
+  closeBtt.innerHTML = '<div></div><div></div>';
   popupWindow.innerHTML = `
-    <div id="comments-popup">
       <div class="img-wrapper">
-          <img src="">IMAGE
+          <img src="${showObj.image.original}">
       </div>
-      <h2 class="show-name">BlackList</h2>
+      <h2 class="show-name">${showObj.name}</h2>
       <div class="details">
-          <span>Genre: Drama</span>
+          <span>Genres  : Drama</span>
           <span>Episodes: 24</span>
           <span>Created by: Jhon Doe</span>
       </div>
-    </div>
   `;
-  document.appendChild(popupWindow);
+  popupWindow.insertBefore(closeBtt, popupWindow.firstChild);
+  closeBtt.addEventListener('click', () => {
+    popupWindow.remove();
+  });
+  document.body.appendChild(popupWindow);
 }
